requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.9.1
bidict==0.22.1
boto3==1.28.58
botocore==1.31.58
CacheControl==0.12.14
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.2.0
cryptography==41.0.4
cycler==0.12.0
dask==2022.10.2
dnspython==2.3.0
docopt==0.6.2
fbmessenger==6.0.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
gast==0.4.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.9.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
jax==0.4.16
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.2
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.5
libclang==16.0.6
locket==1.0.0
magic-filter==1.0.12
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mattermostwrapper==2.2
ml-dtypes==0.3.1
msgpack==1.0.7
multidict==5.2.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
partd==1.4.1
Pillow==10.0.1
pluggy==1.3.0
portalocker==2.8.2
prompt-toolkit==3.0.28
protobuf==4.23.3
psycopg2-binary==2.9.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.1
pyrsistent==0.19.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2022.7.1
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rasa==3.6.10
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.7.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.11.3
sentry-sdk==1.14.0
simple-websocket==0.10.1
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.22.0
SQLAlchemy==1.4.49
structlog==23.1.0
structlog-sentry==2.0.3
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-macos==2.12.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
toolz==0.12.0
tqdm==4.66.1
twilio==8.2.2
typing-utils==0.1.0
typing_extensions==4.8.0
tzlocal==5.0.1
ujson==5.8.0
urllib3==1.26.17
uvloop==0.17.0
wcwidth==0.2.8
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
